requirements.txt
pyyaml>=6.0
loguru>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.0
playwright>=1.40.0
pydantic>=2.0.0
schedule>=1.2.0
flask>=3.0.0
oss2>=2.18.0
gspread>=5.12.0
oauth2client>=4.1.3
